requirements.txt
playwright==1.40.0
pandas==2.0.3
python-dotenv==1.0.0
fake-useragent==1.3.0
requests==2.31.0